fix(users): keep picker unit tests free of the generated api client
The new UserSearch tests imported the component module, whose runtime
import of the generated react-query client broke vitest in CI, where the
client sources are not generated for the lib's test target. Move the pure
helpers (groupUsersByLetter, useDebouncedValue) into UserSearch.utils.ts
with a type-only client import so the tests stay hermetic.
